// src/app/model/user.ts
export class User {
    id?: string;
    abnumber: string;
    firstName: string;
    lastName: string;
    displayName: string;
    email: string;
    jobDescription: string;
    initials: string;
    division: string;
    department: string;
    telephoneNumber: string;
    memberOf?: string [];
    token?: string;
}